Fix #383: TypeError if arg[0] is socket object
The copy operation raises a TypeError because the socket is not serializable. Copying the
object should only prevent the handler from changing it so "orig.excpetion(e)" can still be
executed. However, the socket object does not have a "exception" method so it is save to
only hold a reference in orig.

This can be removed once everything is read-only as described in the todo in line 1175